[ftinspect] Fix parent-child relationships in Qt objects tree.
Most Qt objects' parent should be explicitly set to have the object included
in the Qt object tree. This ensures that all objects is properly deleted when
their parents is being deleted.

Note that `QLayout::addWidgets` doesn't take ownership of the widget, so no
auto-parenting is done here! However, `QWidget::setLayout` does set the parent
of the layout automatically to the widget, so no need to specify parents for
layouts.

* src/ftinspect/maingui.cpp: Add `this` for all non-layout sub-widgets.

* src/ftinspect/widgets/custom_widgets.cpp,
  src/ftinspect/widgets/custom_widgets.hpp,
  src/ftinspect/models/ttsettingscomboboxmodel.cpp,
  src/ftinspect/models/ttsettingscomboboxmodel.hpp: Add `parent` as
  constructor parameters and pass it up to the base class constructor.

* src/ftinspect/panels/settingpanel.cpp,
  src/ftinspect/panels/settingpanel.hpp: Both of the above.