[ftinspect] Support infinite panning
Now the `Grid` dynamically resizes (and resizes the outer scene as well) so
one can drag/scroll infinitely on the grid. "Go back to center" is especially
helpful when you get lost.

* src/ftinspect/rendering/grid.hpp, src/ftinspect/rendering/grid.cpp: Add
`updateRect` to dynamically update bounding rect of the Grid and the scene's
rect.

* src/ftinspect/maingui.hpp, src/ftinspect/maingui.cpp: Call
`Grid::updateRect` when necessary. Note that `MainGUI` must manually maintain
the ownership of `gridItem_` pointer to prevent crash.