[ftinspect] WIP: Add initial support to "Continuous View".
This commit adds a tab to the main GUI named "Continuous View",
corresponding to the `ftview` tool. Inside the tab is a glyph canvas and
another tab view. The tab view is used to switch modes and input settings.

The rendering canvas is *not* a `QGraphicsView`, but a simple `QWidget` with
a custom `paintEvent`. This is because we don't need features like panning
and zooming, so using `QGraphicsView` will only add complexity (e.g.
transforming coordination systems).

Code translating FreeType outline to the origin for rendering is extracted
as a helper function in a new "render utils" file.

The font size selector (size, unit, DPI) in the original `SingularTab` is
extracted as a new class `FontSizeSelector`.

The "graphics defaults", which are graphics objects like color palettes and
`QPen`s, are all extracted into a new singleton struct `GraphicsDefaults`,
so every component can easily obtain them.

Currently there're a lot of settings, status variables and parameters during
the rendering process. Will write some documentation clearifing them all.

WIP: Only "all glyphs" view is implemented.
TODO: Fancy, Stroked, Text String mode; LCD Anti-Aliasing; Bitmap font;

The rendering code is adopted from singular rendering, which is different
from the code in `ftcommon.c` from the legacy tools. In the future we may
need to re-implement this with code from the latter.

* src/ftinspect/rendering/glyphcontinuous.cpp,
  src/ftinspect/rendering/glyphcontinuous.hpp: New file: the actual
  "rendering canvas" in the "Continuous View". The code was a combination of
  `ftcommon` and the singular view's render views. Mouse scrolling is
  supported for navigation and resizing.

* src/ftinspect/panels/continuous.cpp, src/ftinspect/panels/continuous.hpp:
  New files: the content of the "Continuous View" tab is a `ContinuousTab`.
  It current only support one single mode and submode.

  Settings about "All Glyphs" mode is placed in `ContinousAllGlyphsTab`.

* src/ftinspect/engine/engine.cpp, src/ftinspect/engine/engine.hpp: Add
  CharMap handling code, especially `CharMapInfo` struct. `CharMapInfo` can
  provide the name and max index for the charmap. Vector of `CharMapInfo`
  for current font is obtained via `Engine::currentFontCharMaps`.

  CharMap mapping is done via `CharMapInfo::glyphIndexFromCharCode`, with
  the added `FTC_CMapCache`.

  Add more actions to support continuous view. `reloadFont` and
  `loadGlyphWithoutUpdate` can load fonts and glyphs with minimal overhead.
  `loadGlyphWithoutUpdate` uses `FTC_ImageCache_Lookup`, so the
  `FTC_ImageTypeRec` object is cached.

* src/ftinspect/widgets/fontsizeselector.cpp,
  src/ftinspect/widgets/fontsizeselector.hpp: New file, include the font
  size setting part. It can directly apply settings to the engine via
  `applyToEngine` function, and can handle wheel event from event object or
  by steps count.

* src/ftinspect/widgets/glyphindexselector.cpp,
  src/ftinspect/widgets/glyphindexselector.hpp: Combine `setMax` and
  `setMin` functions, and reduce unnecessary signals by blocking them.

  Rename `getCurrentIndex` to `currentIndex`.

  Support custom number rendering (for Unicode charcode formatting). Adjust
  layout and disabled wrapping on the spinbox.

* src/ftinspect/rendering/graphicsdefault.cpp,
  src/ftinspect/rendering/graphicsdefault.hpp: New file, including graphics
  initializing code.

* src/ftinspect/rendering/renderutils.cpp,
  src/ftinspect/rendering/renderutils.hpp: New file, including outline
  translating code.

* src/ftinspect/rendering/glyphbitmap.cpp: Move out outline tranlating code.

* src/ftinspect/panels/singular.cpp, src/ftinspect/panels/singular.hpp: Move
  out the font size selector and graphics defaults. Minor change due to
  change in `GlyphIndexSelector`.

* src/ftinspect/maingui.cpp, src/ftinspect/maingui.hpp: Add "Continuous
  View" to tab view.

* src/ftinspect/CMakeLists.txt, src/ftinspect/meson.build: Updated.